refactor: make driver naming consistent
