+++ src/main/java/edu/iastate/ece/sd/sdmay2126/runner/selenium/driver/SeleniumDriverConfiguration.java
@@ -4,15 +4,15 @@ package edu.iastate.ece.sd.sdmay2126.runner.selenium.driver;
  * Specifies web-driver configuration for Selenium automation.
  */
 public class SeleniumDriverConfiguration {
-    private SeleniumWebDrivers driverType;
+    private SeleniumDrivers driverType;
     private String driverLocation;
 
-    public SeleniumDriverConfiguration(SeleniumWebDrivers driverType, String driverLocation) {
+    public SeleniumDriverConfiguration(SeleniumDrivers driverType, String driverLocation) {
         this.driverType = driverType;
         this.driverLocation = driverLocation;
     }
 
-    public SeleniumWebDrivers getDriverType() {
+    public SeleniumDrivers getDriverType() {
         return driverType;
     }
