+++ src/main/java/edu/iastate/ece/sd/sdmay2126/runner/selenium/driver/SeleniumDrivers.java
@@ -3,7 +3,7 @@ package edu.iastate.ece.sd.sdmay2126.runner.selenium.driver;
 /**
  * Indicates which Selenium WebDrive and web browser to use.
  */
-public enum SeleniumWebDrivers {
+public enum SeleniumDrivers {
     FIREFOX,
     CHROME
 }
